avalok-agent/avalok-ai/requirements.txt
google-adk[database]==0.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
google-cloud-firestore==2.13.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
firebase-admin==6.3.0